Rename //toolchain/common to base
