+++ toolchain/lexer/BUILD
@@ -182,7 +182,7 @@ cc_library(
         "//common:check",
         "//common:ostream",
         "//common:string_helpers",
-        "//toolchain/common:index_base",
+        "//toolchain/base:index_base",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/source:source_buffer",
         "@llvm-project//llvm:Support",
@@ -210,7 +210,7 @@ cc_test(
         ":tokenized_buffer_test_helpers",
         "//testing/util:gtest_main",
         "//testing/util:test_raw_ostream",
-        "//toolchain/common:yaml_test_helpers",
+        "//toolchain/base:yaml_test_helpers",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:mocks",
         "@com_google_googletest//:gtest",
