+++ toolchain/lexer/tokenized_buffer_test.cpp
@@ -16,7 +16,7 @@
 #include "llvm/Support/SourceMgr.h"
 #include "llvm/Support/raw_ostream.h"
 #include "testing/util/test_raw_ostream.h"
-#include "toolchain/common/yaml_test_helpers.h"
+#include "toolchain/base/yaml_test_helpers.h"
 #include "toolchain/diagnostics/diagnostic_emitter.h"
 #include "toolchain/diagnostics/mocks.h"
 #include "toolchain/lexer/tokenized_buffer_test_helpers.h"
