+++ toolchain/parser/BUILD
@@ -47,7 +47,7 @@ cc_library(
         "//common:error",
         "//common:ostream",
         "//common:vlog",
-        "//toolchain/common:pretty_stack_trace_function",
+        "//toolchain/base:pretty_stack_trace_function",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/lexer:token_kind",
         "//toolchain/lexer:tokenized_buffer",
@@ -65,7 +65,7 @@ cc_test(
         "//common:ostream",
         "//testing/util:gtest_main",
         "//testing/util:test_raw_ostream",
-        "//toolchain/common:yaml_test_helpers",
+        "//toolchain/base:yaml_test_helpers",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:mocks",
         "//toolchain/lexer:tokenized_buffer",
