+++ toolchain/semantics/BUILD
@@ -36,7 +36,7 @@ cc_library(
         ":semantics_node_kind",
         "//common:check",
         "//common:ostream",
-        "//toolchain/common:index_base",
+        "//toolchain/base:index_base",
         "//toolchain/parser:parse_tree",
         "@llvm-project//llvm:Support",
     ],
@@ -83,7 +83,7 @@ cc_library(
         "//common:check",
         "//common:ostream",
         "//common:vlog",
-        "//toolchain/common:pretty_stack_trace_function",
+        "//toolchain/base:pretty_stack_trace_function",
         "//toolchain/diagnostics:diagnostic_kind",
         "//toolchain/lexer:numeric_literal",
         "//toolchain/lexer:token_kind",
@@ -119,7 +119,7 @@ cc_test(
     deps = [
         "//testing/util:gtest_main",
         "//testing/util:test_raw_ostream",
-        "//toolchain/common:yaml_test_helpers",
+        "//toolchain/base:yaml_test_helpers",
         "//toolchain/driver",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
