+++ toolchain/semantics/semantics_ir_test.cpp
@@ -11,7 +11,7 @@
 #include "llvm/Support/VirtualFileSystem.h"
 #include "llvm/Support/raw_ostream.h"
 #include "testing/util/test_raw_ostream.h"
-#include "toolchain/common/yaml_test_helpers.h"
+#include "toolchain/base/yaml_test_helpers.h"
 #include "toolchain/driver/driver.h"
 
 namespace Carbon::Testing {
